feat(loan): implement create guarantor screen

+++ core/network/src/commonMain/kotlin/com/mifos/core/network/model/GetMakerCheckerResponse.kt
@@ -64,7 +64,7 @@ data class GetMakerCheckerResponse(
 
     val loanAccountNo: String? = null,
 
-    val loanId: Long? = null,
+    val loanId: Int? = null,
 
     @Contextual
     val madeOnDate: LocalDateTime? = null,
